refacor + chore: create auth system + change codebase to clean architecture

+++ src/main/java/com/bmstu/lab/application/service/CalculateCpiCategoryService.java
@@ -5,10 +5,13 @@ import com.bmstu.lab.infrastructure.persistence.entity.CalculateCpi;
 import com.bmstu.lab.infrastructure.persistence.entity.CalculateCpiCategory;
 import com.bmstu.lab.infrastructure.persistence.entity.CalculateCpiCategoryId;
 import com.bmstu.lab.infrastructure.persistence.entity.Category;
+import com.bmstu.lab.infrastructure.persistence.entity.User;
+import com.bmstu.lab.infrastructure.persistence.enums.CalculateCpiStatus;
 import com.bmstu.lab.infrastructure.persistence.repository.CalculateCpiCategoryRepository;
 import jakarta.transaction.Transactional;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -30,18 +33,41 @@ public class CalculateCpiCategoryService {
   }
 
   @Transactional
-  public void delete(CalculateCpi cpi, Category category) {
+  public void delete(CalculateCpi cpi, Category category, String username) {
+    if (!canModify(cpi, username)) {
+      throw new AccessDeniedException("Недостаточно прав для удаления категории из расчёта CPI");
+    }
+
     calculateCpiCategoryRepository.deleteByCalculateCpiAndCategory(cpi, category);
   }
 
-  public CalculateCpiCategory update(CalculateCpi cpi, Category category, Double userSpent) {
+  @Transactional
+  public CalculateCpiCategory update(
+      CalculateCpi cpi, Category category, Double userSpent, String username) {
+
+    if (!canModify(cpi, username)) {
+      throw new AccessDeniedException("Недостаточно прав для изменения категории расчёта CPI");
+    }
+
     CalculateCpiCategory entity =
         calculateCpiCategoryRepository
             .findById(new CalculateCpiCategoryId(cpi.getId(), category.getId()))
             .orElseThrow(() -> new CalculateCpiCategoryNotFoundException("Запись не найдена"));
 
-    if (userSpent != null) entity.setUserSpent(userSpent);
+    if (userSpent != null) {
+      entity.setUserSpent(userSpent);
+    }
 
     return calculateCpiCategoryRepository.save(entity);
+  }
+
+  private boolean canModify(CalculateCpi cpi, String username) {
+    User creator = cpi.getCreator();
+
+    boolean isCreator = creator.getUsername().equals(username);
+    boolean isModerator = creator.isModerator();
+    boolean isDraft = cpi.getStatus() == CalculateCpiStatus.DRAFT;
+
+    return (isCreator && isDraft) || isModerator;
   }
 }
